refactor: share common functions for different git sources
