+++ src/utils/dispatch-update.ts
@@ -1,13 +1,14 @@
 import vscode from 'vscode';
 import { updateFileSystem } from '../sources/filesystem';
-import { updateForgejo } from '../sources/forgejo';
-import { updateGitHub } from '../sources/github';
+import * as Forgejo from '../sources/forgejo';
+import * as Git from '../sources/git';
+import * as GitHub from '../sources/github';
 import { updateMarketplace } from '../sources/marketplace';
 import { Source, UpdateResult } from './types';
 
 export async function dispatchUpdate(extensionName: string, currentVersion: string, source: Source, temporaryDir: string, debugChannel: vscode.OutputChannel | undefined): Promise<UpdateResult> {
 	if(source === 'github') {
-		return updateGitHub(extensionName, currentVersion, undefined, temporaryDir, debugChannel);
+		return Git.update(extensionName, currentVersion, undefined, GitHub, temporaryDir, debugChannel);
 	}
 
 	if(source.type === 'file') {
@@ -15,11 +16,11 @@ export async function dispatchUpdate(extensionName: string, currentVersion: stri
 	}
 
 	if(source.type === 'forgejo') {
-		return updateForgejo(extensionName, currentVersion, source, temporaryDir, debugChannel);
+		return Git.update(extensionName, currentVersion, source, Forgejo, temporaryDir, debugChannel);
 	}
 
 	if(source.type === 'github') {
-		return updateGitHub(extensionName, currentVersion, source, temporaryDir, debugChannel);
+		return Git.update(extensionName, currentVersion, source, GitHub, temporaryDir, debugChannel);
 	}
 
 	if(source.type === 'marketplace') {
